Addition of Third Panel and Player Territory Pick
Addition of a third panel on the right hand side of the map which we will modify to allow textual commands from the user. We will then modify this panel to allow the user to pick their preferred territory (and the neutrals territory) and allow the user to enter other commands for the program.